requirement.txt
torch==2.4.1
torchvision==0.19.1
torchaudio==2.4.1
numpy==2.1.1
pandas==2.2.2
scikit-learn==1.5.2
scikit-image==0.25.2
Pillow==10.3.0
tqdm==4.67.1